+++ src/pages/myPage/ui/components/index.ts
@@ -1,3 +1,4 @@
 export { default as Divider } from './Divider'
 export { default as PlaylistGrid } from './PlaylistGrid'
+export { default as SubHeader } from './SubHeader'
 export { default as UserProfile } from './UserProfile'
